get and delete for jobs
