+++ client/package.json
@@ -14,6 +14,7 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "axios": "^1.6.5",
     "better-react-mathjax": "^2.0.3",
     "html2canvas": "^1.4.1",
     "jspdf": "^2.5.1",
@@ -21,7 +22,7 @@
     "react-dom": "^17.0.0 || ^18.0.0",
     "react-redux": "^9.1.0",
     "react-router-dom": "^6.21.1",
-    "react-scripts": "^3.0.1",
+    "react-scripts": "^5.0.1",
     "react-zoom-pan-pinch": "^3.3.0",
     "web-vitals": "^2.1.4"
   },
