+++ client/src/App.js
@@ -14,27 +14,16 @@ function App() {
   return (
     <Provider store={store}>
       <Router>
-        <Box
-          mt={4}
-          mb={4}
-          textAlign="center"
-          display="flex"
-          alignItems="center"
-          justifyContent="center"
-        >
+        <Box mt={4} mb={4} textAlign="center" display="flex" alignItems="center" justifyContent="center">
           {/* Logo and Title */}
-          <img
-            src={logo}
-            alt="Logo"
-            style={{ width: "55px", marginRight: "10px", marginBottom: "10px" }}
-          />
+          <img src={logo} alt="Logo" style={{ width: "55px", marginRight: "10px", marginBottom: "10px" }} />
           <Typography variant="h5">TAILOR MY RESUME</Typography>
         </Box>
         <Routes>
           <Route path="/home" element={<Home />} />
           <Route path="/resume" element={<BuildResume />} />
           <Route path="/job" element={<JobDetail />} />
-          <Route path="/" element={<BuildResume />} />
+          <Route path="/" element={<Home />} />
         </Routes>
       </Router>
     </Provider>
