+++ index.js
@@ -2,7 +2,7 @@ import express from "express";
 import mongoose from "mongoose";
 import dotenv from "dotenv";
 import routes from "./src/routes/index.js";
-
+import cors from "cors";
 dotenv.config();
 const app = express();
 const PORT = process.env.PORT || 3030;
@@ -28,6 +28,7 @@ db.once("open", () => {
   console.log("Connected to MongoDB");
 });
 
+app.use(cors());
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
